Add spacing for each dimension (#11)
This adds a new widget to the main plugin widget where each row shows
and allow control of each dimension's space (e.g. pixel size). This
deviates from the initial designs in an effort to accommodate napari's
general n-dimensional layer approach, but gets closer to it and is
similar in spirit.

![Screenshot 2023-02-03 at 2 25 08
PM](https://user-images.githubusercontent.com/2608297/216722987-77df4758-ee73-42cd-b69d-08a431a45ccc.png)

I chose to use a spinbox to indicate this is a numeric value and to
allow easy modification and validation of values manually edited. We
could consider a line edit with a custom positive double validator later
though, as spin boxes can be a little finicky.

I also removed some of the early prototype code and simplified the
read-only attributes.

Closes #4.