Toggle between full and editable metadata (#12)
This adds a checkable button to the metadata widget to control if only
the editable or the full metadata are shown. By default, the button is
unchecked and only the editable metadata are shown. When the button is
checked, the full metadata are shown.

![Screenshot 2023-02-03 at 3 46 18
PM](https://user-images.githubusercontent.com/2608297/216731227-5669c5f5-c96e-4514-8b72-5df38485f58c.png)

Closes #5