Add a writer contribution that uses plugin specific extra metadata (#18)
This adds a writer contribution that uses the data stored in the extra
layer metadata defined by this plugin as well as the standard napari
layer attributes (e.g. `scale`).

I also removed the current read/write buttons and the associated code,
since those have been broken for a little while.

There's an opportunity to upstream some of this work to the official
`napari_ome_zarr` plugin, which currently lacks a writer. By
upstreaming, we'd add functionality to that more widely known plugin and
also simplify the implementation here.

I added tests to cover 2d, 3d, and multi-scale images.

Closes #9, #15