Add a reader contribution that supports ome-zarr (#19)
This adds a reader contribution that vendors the reader from
[napari-ome-zarr](https://github.com/ome/napari-ome-zarr) and makes some
modifications to it so that the extra-napari metadata support this
plugin enables (e.g. axes) can be read from the ome-zarr directory. It
also fixes some potential bugs in napari-ome-zarr I found while writing
tests.

I also fixed some integration bugs found while doing an edit-write-read
loop. I'm sure there are more to find and fix, but this gets a few of
them.