Store original attributes and allow restoring them from the widget (#20)
This stores the original values of attributes shown in the widget in the
data model of this plugin, then also adds a button that causes those to
be restore to their original/default values.

A natural place to populate those original values is at read time, so we
do that in our reader contribution. For layers that were read by other
plugins, things become harder. Currently, we populate the extra and
original metadata attributes only when a layer is selected and our
widget is active, so these might not be the values when the layer was
read. Alternatively, we could populate those when a layer is added and
our widget is active, though that still relies on our widget being
active.

Closes #13