Use a stacked widget to switch between editable and read-only metadata (#30)
There's a lot of code duplication here in order to get read-only
versions of the widgets. But as I'm planning to do some
redesign/refactoring anyway, I think it's worth getting this in and
working from there.

This also hides the save and restore default button appropriately when
viewing the read-only widget.