Only enable restore defaults when metadata is different to original (#37)
This updates the enabled state of the restore defaults button anytime
anything the editable metadata widget cares about changes. Some of these
changes are connected to widget signals, while others are connected to
napari model events. Would be nice to have a central place (e.g. the
plugin model) to observe for these changes, but this should be good
enough for now.

Closes #24.