Show axis and scale bar when showing widget (#39)
This shows the axes and scale bar overlay on the canvas whenever the
widget is shown. This adds convenience since a large part of this widget
is related to what is shown on those. This might be a minor annoyance to
some users who want these to be non-visible, but they can still use the
napari built-in controls to turn them off.

Closes #27