Add translation control for each dimension (#42)
This adds similar spinboxes to control the translation/offset of each
dimension of a layer. It places them next to the scale spinboxes, though
the layout may be improved in the future. There is also some refactoring
here to improve code sharing a little, but much more is still needed.

Closes #10