Use compact double line edit instead of spin box (#55)
This uses a custom line edit that mostly behaves like a double spinbox,
except that it does not round the current value to a specific number of
decimals. This makes for less frustrating interaction when specifying
certain values and also makes things more compact by default. We do lose
the interactivity of the spin box (which could be added back with some
effort), but this widget is about metadata, not about manual alignment
when that might be more relevant.

I also simplified some module and class names.