File size widget (#21)
This PR adds a file size widget to the read only attributes. For local
files (those with `layer.source.path`), it will collect the actual size
on-disk. For remote files (those whose `layer.source.path` is None or
points to a URL), it will collect the in-memory size and display "in
memory" next to the file size.

For local zarr files, it should be possible to determine the on-disk
size, but since the `napari-ome-zarr` plugin doesnt store the path, we
cant get it. Therefore, local zarr files are given as in memory size.

The floats displayed are rounded to 2 decimal places. 

A reasonable transformation of the size is also done such that it
displays "12.34 KB" or "12.34 GB" as appropriate.